# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação do gerenciamento de biblioteca de jogos 🚀",
        "description": "Esta é a documentação gerada pelo Swagger para a minha API de gerenciamento de biblioteca de jogos.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Lista todos os jogos",
                "operationId": "93736f40a135111a310e4b25ad61f30b",
                "responses": {
                    "200": {
                        "description": "Lista de jogos retornada com sucesso"
                    },
                    "400": {
                        "description": "Jogos não encontrados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Criar um novo jogo",
                "operationId": "b8b9557618eb0f7b3650bdd4ab3a6487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "publisher",
                                    "platforms"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Hollow Knight"
                                    },
                                    "publisher": {
                                        "type": "string",
                                        "example": "Team Cherry"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "example": "2021-02-20"
                                    },
                                    "platforms": {
                                        "type": "string",
                                        "example": "Nintendo Switch, PS4, PS5, Xbox, PC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jogo criado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "500": {
                        "description": "Erro ao criar jogo"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar todos os usuários",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso"
                    },
                    "400": {
                        "description": "Usuários não encontrados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar um novo usuário",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "500": {
                        "description": "Erro ao criar usuário"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar usuário pelo Id",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrados"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar o usuário",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "_method"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "put"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário atualizado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Usuário não encontrados"
                    },
                    "500": {
                        "description": "Erro ao atualizar usuário"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deletar usuário",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "delete"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Usuário não encontrados"
                    },
                    "500": {
                        "description": "Erro ao deletar usuário"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Games",
            "description": "Games"
        },
        {
            "name": "Usuários",
            "description": "Usuários"
        }
    ]
}